Split HurlParser parsing functions to sub files.
