+++ hurl-core/src/main/kotlin/com/orange/ccmd/hurl/core/parser/Parser.kt
@@ -17,22 +17,9 @@
  *
  */
 
-package com.orange.ccmd.hurl.core.ast
+package com.orange.ccmd.hurl.core.parser
 
 import com.orange.ccmd.hurl.core.utils.slice
-
-data class Position(
-    var offset: Int = 0, // Offset in code points, starting at 0
-    var line: Int = 1,   // Line number in code points, starting at 1
-    var column: Int = 1  // Column number in code points, starting at 1
-) {
-    val text: String
-        get() = "[$line:$column]"
-
-    companion object {
-        val zero: Position = Position()
-    }
-}
 
 typealias ParseFunc<T> = Parser.() -> T?
